Rename: recoil navState 파일 분리
